+++ kkm/gui/dialogs/__init__.py
@@ -19,6 +19,7 @@
 from kkm.core import Capability
 from .config_dialog import ConfigDialog
 from .config_backup_dialog import ConfigBackupDialog
+from .factory_reset_dialog import FactoryResetDialog
 from .firmware_dialog import FirmwareDialog
 from .user_dialog import UserDialog
 from .onvif_dialog import OnvifDialog
@@ -29,11 +30,12 @@ from .ip_dialog import IpDialog
 ACTION_DIALOGS = {
     Capability.CONFIG: ConfigDialog,
     Capability.CONFIG_BACKUP: ConfigBackupDialog,
+    Capability.FACTORY_RESET: FactoryResetDialog,
     Capability.FIRMWARE: FirmwareDialog,
     Capability.USERS: UserDialog,
     Capability.ONVIF_USERS: OnvifDialog,
     Capability.SET_IP: IpDialog,
 }
 
-__all__ = ["ACTION_DIALOGS", "ConfigDialog", "ConfigBackupDialog", "FirmwareDialog",
-           "UserDialog", "OnvifDialog", "IpDialog"]
+__all__ = ["ACTION_DIALOGS", "ConfigDialog", "ConfigBackupDialog", "FactoryResetDialog",
+           "FirmwareDialog", "UserDialog", "OnvifDialog", "IpDialog"]